backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic[email]==2.9.0
motor==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
websockets==13.1
email-validator>=2.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

